enable resource loading for android project
- added an IResourceLoader interface with a since load
loadJsonInputStreams method
- added an AssetResourceLoader in the android module to load files from
the android asset folder
- updated ResourceLoader to implement the IResourceLoader interface
- updated ResourceLoader with an init method that must be called with
an IResourceLoader impl before any resources can be loaded.  The
android project will call it with an instance of the
AssetResourceLoader and the java app project will call it with an
instance of the ResourceLoader instance
- updated CardCatalogue with usage of ResourceLoader.getInstance() and
removed usage of the Path class since it not available on Android.
Replaced it with usage of File instead
- updated ApplicationFacade to initialize the ResourceLoader with a new
instance